refactor: Centralize scroll utility functions (refs #21)
